Update patch set 1
Patch Set 1:

> https://bugs.eclipse.org/bugs/show_bug.cgi?id=353460
> But now the bug is the other way around -> No ASCII processes, which probably I thought was less bad than no Unicode?
> Who sets win32.ascii property? If a big majority of setups will have Unicode by default anyway, it sounds like it will mostly be like the old behavior so I think it's OK. We can also put the win32.ascii property in the bug details and release notes.

Hm. I apparently never submitted my reply to your questions.

The w32.ascii property is up to the user that launches the JVM to set, but there is no need to do it. In the case of Eclipse, I do not think we should set it.
I tried to verify that w32.ascii=true would actually work but I found out that there are several parts of the JNA code base that does not properly handle the ANSI version of the win32 API. (I have an ongoing discussion with the JNA developers about fixing this).

To my knowledge, both the ANSI version and the UNICODE version of the win32 API lists all processes in the system. I tried to compile a dummy hello world application using MinGW and I can see the processes with both the ANSI and the UNICODE version of the win32 (tested both with and without -D_UNICODE and -DUNICODE compiler flags).
Do you have a clear use-case where the processes are missing with listtasks.exe that I can try?

Patch-set: 1