Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I think this is great - a few lines of Java using JNA (even less when JNA 5.6 comes out) is great.
> > 
> > Marc-Andre, what is your vote on this? Do we need to test the unicode for ansi process names, or is this good as is?
> 
> I'll give the patch a test tomorrow. If it's more or less the status-quo (Unicode process names only?) then I think it's OK. It's not like ansi should become more popular and we have increasing needs to polish Windows support for this corner case.

Yes, for 64-bit Windows NT based systems, the process list provided by listtasks.exe and this implementation should be identical.

I did not port the support for listing 16-bit applications or the support for non-Windows NT based systems. I think it would be straight forward to port support for these too, but I have no way to verify it. Is this required to merge this changeset?

Patch-set: 1