Update patch set 2
Patch Set 2:

Hi Alex, this patch addresses the gdb issue.  It appears that running the gdb on the host is the issue so simply not running gdb under flatpak-spawn works fine and the terminal I/O is correctly routed to the Eclipse Console.  Feel free to play with it a bit and let me know if you find any issues.

Patch-set: 2
Reviewer: Gerrit User 726 <726@97ee7c02-f12f-4043-b43e-dea463d88b31>