Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> @Dirk it looks a bit alien for this repo, all these Bndtools-related identifiers and java nature. I would expect it to be contributed from the Bndtools side to not introduce implicit circular dependency. WDYT?

I actually don't know if Bndtools could contribute an extension that it does not know. I mean, the contribution is using classes from the tm.terminal plugins. Can another plugin that is not aware of those plugins contribute extension based on such classes?

For the JDT Package Explorer also the tm.terminal plugin makes the contribution. That was done some years ago, but I thought it could then also be done for Bndtools. But I agree that it looks strange, even the implicit circular dependency is only a weak one from my experience, as it is only based on an identifier and should be ignored if the identifier is not located in the runtime. At least if my understanding of the extension point mechanism is correct at this place.

In the end I am just a contributer here and wanted to have the context menu entry also in the Bndtools Explorer. The decision has to be made by the committers.

Patch-set: 2