Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > @Dirk it looks a bit alien for this repo, all these Bndtools-related identifiers and java nature. I would expect it to be contributed from the Bndtools side to not introduce implicit circular dependency. WDYT?
> 
> I actually don't know if Bndtools could contribute an extension that it does not know. I mean, the contribution is using classes from the tm.terminal plugins. Can another plugin that is not aware of those plugins contribute extension based on such classes?
> 
> For the JDT Package Explorer also the tm.terminal plugin makes the contribution. That was done some years ago, but I thought it could then also be done for Bndtools. But I agree that it looks strange, even the implicit circular dependency is only a weak one from my experience, as it is only based on an identifier and should be ignored if the identifier is not located in the runtime. At least if my understanding of the extension point mechanism is correct at this place.
> 
> In the end I am just a contributer here and wanted to have the context menu entry also in the Bndtools Explorer. The decision has to be made by the committers.

I don't have strong objections to include this patch. 
Actually I want to discuss this architectural problem with the well known OSGi and RCP expert :) 
So, we have an ability to reference the unknown identifier from the "supplier" side and hope that we will not be panished by "the framework" based on our experience.
At the same time, we don't want to reference the unknown identifier from the "consumer" side because we afraid that "the framework" will reject it.
From my point we should add extension to "consumer" (Bndtools) to know about the new functionality "supplier" (TM terminal), and we should do it explicitely to ensure that functionality will be derived. May be it should be an optional feature, but this will make things much more transparent. But I have no idea where the Bndtools are hosted and would it be accepted there.

Patch-set: 2