Update patch set 2
Patch Set 2:

> Patch Set 2:
> I don't have strong objections to include this patch. 
> Actually I want to discuss this architectural problem with the well known OSGi and RCP expert :) 

If you want to discuss this with some well known OSGi and RCP expert, you should add that person as a reviewer. ;-)

> So, we have an ability to reference the unknown identifier from the "supplier" side and hope that we will not be panished by "the framework" based on our experience.

AFAIK the context menu contribution from the supplier side works, because we reference the context menu to which we want to contribute by ID. If that ID is not found the framework will do nothing as the menu to contribute to is not found.

> At the same time, we don't want to reference the unknown identifier from the "consumer" side because we afraid that "the framework" will reject it.

AFAIK consuming a command from another plugin and make a contribution to the local context menu will not work, because the contribution uses a class, not an ID. So this will already result in an error inside the IDE and probably ClassNotFoundExceptions.

> From my point we should add extension to "consumer" (Bndtools) to know about the new functionality "supplier" (TM terminal), and we should do it explicitely to ensure that functionality will be derived. May be it should be an optional feature, but this will make things much more transparent. But I have no idea where the Bndtools are hosted and would it be accepted there.

Bndtools is located on GitHub and is not setup using PDE. So this could get complicated. When I worked on the Git Bash integration for TM Terminal several years ago I had similar thoughts and did not find any solution on this. The difference is reference by ID (weak type of dependency) to reference by class (strong type of dependency). And IIRC this is also the reason why the java nature and JDT Package Explorer reference already exist this way and are not part of JDT.

But to be honest, my knowledge of the Extension Point mechanism is not deep enought. I am not a big fan of it anymore and use OSGi DS where ever possible. If you think that the weak dependency by ID would also break things from the TM Terminal side, or the strong dependency by class would not break things on the Bndtools side, please let me know. My experience tells me that it does not work the other way round. But I might be wrong.

To don't move the discussion into the wrong direction, in general I agree and would not suggest to introduce such a dependency (weak or strong) on the supplier side. It feels wrong. In a plain service design the supplier (TM Terminal) would contribute the provided command in some registry, then on the consumer side you would receive the command if available (hopefully by an interface) and then do with it whatever you like. If the command would be only a definition it could probably work, but here it is a dynamic contribution for which the class is needed. And I don't know how this could be accomplished without introducing a strong dependency.

Do you maybe know if the whole contribution design in this place probably needs to be changed to make it work the way you suggest? IMO even the E4 Model contributions work in this way that the supplier tells to which model element something should be contributed to. Although in most cases the elements to contribute to are global elements like the toolbar. But here also not the toolbar as the consumer specifies what it wants to contribute, it is the supplier that contributes. Or am I wrong?

Patch-set: 2