Bug 565553 - Improve performance of build command parsers with large number of files
Compile a pattern ahead of time and early return.

The first pattern is used to trim extra characters that are not
contained within the actual option pattern. If this first pattern
doesn't match, there is no point in continuing because it is a superset.
By trying this, it exposed a flaw with the use of this "trimming"
pattern. Since the option pattern is allowed to contain back reference
with numbered groups (e.g. \1), then embedding it within another pattern
affects the group numbers and what that back reference matches.  Only
the GCC parser for quoted options was using numbered group references
and because its generic options pattern does not produce extra
characters for quoted values, then this failed replacement was silently
ignored, without producing any different outcome. Since we now early
return based on that pattern matching, it exposes the problem fully.

In theory, even before this proposed change, another parser might not be
as lucky as the GCC parser and its usage of numbered groups might break
the trimming logic. Therefore, I clarified in the javadoc that using
numbered group back-references in the option pattern is not safe and
fixed the GCC patterns.  This is technically a modification of the API
contract but this class is still marked as experimental and it is also
regarding an edge case that is not high risk.

Times before/after, only counting AbstractOptionParser.parseOption
LLVM before: 4289ms, after: 622ms

Change-Id: Id40fc9a35359c39aea00ba14813ffe6c343158fc
Signed-off-by: Marc-Andre Laperle <malaperle@gmail.com>