Update patch set 2
Patch Set 2:

After some experimentation, I believe the way to handle CDT under flatpak is as follows:

1. wrap all local commands in the ProcessFactory using flatpak-spawn
   (this also handles local terminal which we want to run on the host)
2. create a command launcher for flatpak (similar to container support), but
   have it simply use ProcessFactory for commands
3. have the flatpak command launcher copy headers to the workspace similar to how
   container support does it and handle opening of host headers so edit/build/debug
   are consistent with all header info coming from the host
4. use gdbserver under a local terminal to debug the application and use a flatpak gdb
   to debug similar to how we handle container debugging

This would require a gdb existing in flatpak.  Using gdb on the host won't work for terminal I/O
apps as
currently the CDT sets the inferior tty after starting and the tty names are in a
separate space so /dev/tty1 in flatpak is not accessible to the host inferior and likely
overlaps an existing host terminal.  This is why I was seeing the input/output going to various
terminals I had already open.  This problem does not occur just running the terminal I/O application
directly using flatpak-spawn (input/output works from the Eclipse console).

To get gdbserver working I tried hijacking a new local terminal and change the command
to be /bin/sh -c "gdbserver ...." which then got the flatpak-spawn --host wrapping from
ProcessFactory.  This worked great and I/O went to the terminal.  I was not able to emulate this by simply calling flatpak-spawn to set up the gdbserver but I still need to do some investigation into the local run to try and emulate environment.  An alternative would be to delve into the terminal code to see if would be possible to start a local terminal with an initial command.

Patch-set: 2