Update patch set 3
Patch Set 3:

I've moved the source files from org.eclipse.cdt.core.{linux,macosx,win32} to org.eclipse.cdt.core.native to reduce the number of duplicated files.

I did not update the built libraries in this commit, so the build should fail. The reason is that I would like to verify that it actually fails.

The PTY part for Windows is also to be moved to this o.e.c.core.native, but as I'm not sure if it actually builds and that Jonah is/was working on updating it, I left it for the next iteration.
The PTY part also requires the https://github.com/eclipse-cdt/cdt-infra/pull/33 to be merge as the PTY code is using C++.

@Jonah: Feel free to review this gerrit now 😊

Patch-set: 3
Reviewer: Gerrit User 5071 <5071@97ee7c02-f12f-4043-b43e-dea463d88b31>