Update patch set 8
Patch Set 8:

> Patch Set 8: Verified-1 Code-Review-1
> 
> Build Failed 
> 
> https://ci.eclipse.org/cdt/job/cdt-verify-code-cleanliness-pipeline/1740/ : FAILURE. The code base is not clean. This means one or more of formatting, whitespace, version numbers or other checks appears to have failed. Please see the console output for details and https://github.com/eclipse-cdt/cdt-infra/tree/master/jenkins/pipelines/cdt/verify#cdt-verify-code-cleanliness for documentation on what is checked.

I've rebuilt the libraries several times locally in docker and the tree is clean according to git. When doing the same on Jenkins with the same docker image, there is a difference! How can this be?

Patch-set: 8
Reviewer: Gerrit User 5071 <5071@97ee7c02-f12f-4043-b43e-dea463d88b31>