Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I'm not really sure what to do. I don't really agree with renaming it to reindex. I think configurations should be notified that they are active and it's up to them do what they want, like fire events (of course, doing heavy work synchronously is not good practice for extenders). A specific configuration type could have custom UI to notify of the active config change. Unless there is another mechanism to listen to such "active config" change?
> 
> This sounds like an invitation to discussion :) The problem has very deep roots and I'm not sure that we can resolve it on one turn.
> 1) CDT declared its own "Project Build Configuration" idea and started its own journey mostly outside of the platform support.

James didn't get to finish the platform support in order to fully replace the CDT configs. I guess Doug didn't want to continue that work with the new Core build configuration as there is the reputation of being hard to put things in Platform (less so now!), or just didn't get to address this point.

> 2) The appropriate layer to manage changes like "project active configuration changes" is "org.eclipse.core.resources" package, and all others can trigger it via Platform Resources API and listen for changes.
> 3) UI can ask to trigger the event chain (using Job) and then listen for the result to react (asynchronously). But it should not try to implement its own "gear" to manually control the build configuration: it is just not possible in general case, because one will need to duplicate all the Platform Resources implementation, including its native parts.

That's more or less what I had in mind too.

> 4) We are far from the right way at the moment but at least we can start our turn to the right direction, and for this we need to stop introducing another "gear" for manual control of resources.
> 5) If we don't have the appropriate events in Platform Resources, let's understand what we need and then let's introduce them.

I thought it would be a simple clean-up patch (a follow-up to the mailing list question), as is it collapsing existing API so not introducing anything new.
I don't plan to be involved in much Build development as I find it does not provide as much value (I build outside CDT) and considering the small amount of free time I have. This sounds like a more substantial change (perhaps I'm wrong!). So if this patch doesn't go in the right direction, it sounds to me the status quo is better than this patch and it should be abandoned.

> Isn't it nice to see how "leaf" subsystems brutally manipulate the platform resources state using the code copied from `CBuildConfiguration#setActive(IProgressMonitor monitor)`? So, in fact it is used. But the way it is used is something special.

I don't understand, I still see nothing calling setActive(IProgressMonitor monitor) in those three. You mean it should stay so that people can copy its code?

Patch-set: 1