Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Generally looks good. I just wonder why this change didn't affect the tests in any way.

I tried to explain it a bit in the commit message without ranting too much.
Basically, GCC parser only uses numbered back references for quotes and the initial/generic/overall pattern for options, OPTIONS_PATTERN is written in a way that quoted options will never have the extra source file name after it. Because of this, the "get rid of extra text at the end" part never needs to do anything for these GCC cases so it's OK for the pattern to fail silently to match for the replaceFirst call. So for example of pattern fail: for a pattern -I(")foo\1 it will try to match (-I(")foo\1).* which makes no sense because now group #1 is not the same as the parser expected so it will not match, therefore nothing gets replaced.
But another parser using numbered back references in combination with an alternative OPTIONS_PATTERN that allows source file names to go through will not be so lucky and would see the source file name not trimmed properly because of this replaceFirst pattern failing to match.
The optimization here fully exposes the issue because we now don't continue if the replaceFirst pattern doesn't match so what used to fail silently without consequence for some parsers (GCC) now has more consequence (the parser is completely skipped!).

So with this commit, I'm proposing that the optimization alone makes an existing flaw worse and tweaking the javadoc can help prevent both the original flaw and also make the optimization possible.
I will split this into another commit/review before this commit (chain) to demonstrate the original flaw in a unit test (which I think has value either way), fix the javadoc and GCC parser. Then this one will only contain the optimization.

Patch-set: 5