Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> More investigation needed. Torbjörn - do you have the info/computer needed to identify the issue, or would you like me to finish this up?

It would be a lot easier for me if you did it, but at the same time, you have a lot of other tasks on your plate. If possible, can you give it a test try to remove/comment out line 336-340 as I mention in the reply and see if it works then?

Patch-set: 1