Update patch set 4
Patch Set 4: Verified+1

(1 comment)

I have tested on Windows and it looks ok.

There are dozens of code paths, and I have not been able to verify each one though.

Jantje, please +1 this change if you are happy for it to be merged. I have about a week left to get this in in time for 2020-09 M3, and about 3 weeks until RC2 (final release candidate).

Once Jantje and CI +1 this change it is good to go.

Patch-set: 4
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=+1