Update patch set 1
Patch Set 1:

> Yes! I really appreciate the logging instead of the all to often commented out logging that is prevalent in CDT today.

Yes, using the tracing framework is really powerful to get useful data out of a broken system without needing to install another build just to get the needed data from the system. I just wish that it was more obvious to Eclipse developers that it exists and how to use it.

Patch-set: 1