Update patch set 1
Patch Set 1:

Hello Mat,

Jonah said that you may be able to help us out and run a smoke test on PowerPC(ppc64le).
Can you please try https://ci.eclipse.org/cdt/job/cdt-verify-test-cdt-other-pipeline/2276/artifact/releng/org.eclipse.cdt.repo/target/repository/ on a PowerPC and verify that you can start a local terminal and that you can build a hello world application?

If the above 2 operations succeed, I think we can conclude that the libpty.so and libspawner.so is working fine on PowerPC.

Regarding libserial.so; If the above test is successful, I think it's safe to assume that using the same compiler and the options from previous builds would be safe for the serial library too.
If patchset 2 has finished building when you read this Mat, please use that repository instead as it also includes the serial library. If so, just opening the "Launch terminal" UI and selecting "Serial terminal" should be enough to trigger the library to be loaded. If there is no stack trace written to stderr when doing so should mean that the library loaded just fine.

Last but not least, big thanks for helping us run these smoke tests!

Patch-set: 1