Update patch set 10
Patch Set 10:

Hi Mat, forgot to include you on this.  These are changes for using CDT with Eclipse flatpak.  Tools are run on the host and headers are copied from the host so gcc, gdb, glibc, are not be needed in the flatpak.  Please let me know if I have missed something.

Patch-set: 10
CC: Gerrit User 888 <888@97ee7c02-f12f-4043-b43e-dea463d88b31>