Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2: Verified+1 Code-Review+2
> > > > 
> > > > PS2 was just rebuilt because I could not submit PS1 as fast-forward.
> > > 
> > > I really hope someone comes up with a better way of handling the timestamp in the dll issue.
> > > 
> > > Maybe we should hard code it to some arbitrary time in the past. When I first looked it up I found that we couldn't simply hard code timestamp to 0 or so because that causes "bad things". Can't remember what now though.
> > > 
> > > Anyway, it would be nice to not have to use the GIT_COMMITTER_DATE....
> > 
> > Just a wild thought here. What would happen if we use 
> > ```
> > SOURCE_DATE_EPOCH=$(shell git log -1 --pretty=format:%at -- .)
> > ``` 
> > instead? This would use the author time instead and this, you could amend as many times as need without having the conflict. The author time is also persistent when fast forwarding (something that commit time is not).
> > Would this be a better way to handle this or should we just hard-code it to some time in the past?
> 
> Originally I had considered it as author time when I was developing it. On revisiting the question I now think I was overthinking it in the past and I can't remember what good reason I went with committer time (And apparently didn't write it down :-(
> 
> Please make the change, and if the build is clean I can't see a downside now.
> 
> I don't think it should be hard coded, in the past (another non-CDT project) it was super useful to have the timestamp in the DLL to track back its source. Of course, other platforms don't have timestamps at all in the .so, so also probably me overthinking it.

The only downside to %at compared to %ct is that a later timestamp does not have to be a more recent change. I'm not sure how important that is...
Reading up on the documentation of the SOURCE_DATE_EPOCH, I can't find anything about what it's used to in the .dll and .exe files. Googling for a bit turned up https://social.technet.microsoft.com/Forums/en-US/ce2ae265-699f-483b-9890-20ec818cc34a/windows-system-dll-pe-timestamps-are-randomized-after-recent-update?forum=win10itprogeneral
Based on the last 2 comments on that thread, it appears that Microsoft has moved away from this field being a timestamp to rather being a hash of the input. Here is another link that supports that claim, https://devblogs.microsoft.com/oldnewthing/20180103-00/?p=97705 (I think this is the new link for the last post from the link above).

Maybe we should just do that? Hash the source files and use the hash as a 32-bit integer?

Patch-set: 2