Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review+1
> 
> This looks great. I want to have a quick spin of it in Windows before I merge it. But having something that does not have to be fast-forwarded is a positive step.
> 
> Can you also update README.md in root with new instructions (i.e. remove all the amend commit nonsense).

There is one downside of this change and that is that you will have to have python3 on the platform that you are using to build. In docker that's not an issue, but what about building on Windows? I guess the same thing could be scripted in bash, but it will be a much larger script.

@Jonah: If the build for patch set 2 fails too, can you have a look at why the files differ? I can't seam to find the in the artifacts list and I get the same build result each time here locally.

Patch-set: 2