Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+2
> As Marc-Andre had a look earlier I am going to give him a day or two to throw in any final thoughts before I submit this.

Since the launch files can be shared, doesn't it make more sense to pretty format it to help merge/conflict-resolution? I've never used launch files with the combination of the memory browser, so you guys know best.

Otherwise, I have a pending patch to add an argument to enable/disable pretty formatting when calling "serializeXml(Document doc, URI uriLocation, String lineSeparator)", but instead I can replicate the pattern you applied to "toString" and introduce "serializeXml(Document doc, URI uriLocation)" that will not pretty-print. So I think it'll be OK. In case you are wondering, it's for *.language.settings.xml files when they are stored in the workspace plugin state location since they are not meant to be shared and it's a performance issue on large projects with per-file language settings.

Patch-set: 5