Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5: Code-Review+2
> > As Marc-Andre had a look earlier I am going to give him a day or two to throw in any final thoughts before I submit this.
> 
> Since the launch files can be shared, doesn't it make more sense to pretty format it to help merge/conflict-resolution? I've never used launch files with the combination of the memory browser, so you guys know best.

Yes indeed, but in this particular case, it's the value of the value of an attribute in the launch configuration. There is no point in pretty printing the value part as it will be encoded away anyway. Here is an example without pretty printing:
```
<stringAttribute key="org.eclipse.dsf.launch.MEMORY_BLOCKS" value="&lt;?xml version=&quot;1.0&quot; encoding=&quot;UTF-8&quot; standalone=&quot;no&quot;?&gt;&lt;memoryBlockExpressionList context=&quot;reserved-for-future-use&quot;&gt;&lt;gdbmemoryBlockExpression address=&quot;3735928559&quot; label=&quot;0xdeadbeef&quot;/&gt;&lt;gdbmemoryBlockExpression address=&quot;59774856946&quot; label=&quot;0xdeadbeef2&quot;/&gt;&lt;/memoryBlockExpressionList&gt;"/>
```

I don't have an example at hand where the encoding happens, but all \n and \r will be replaced with numeric XML references.
I don't think it makes it any easier to read if there are even more "nonsense" in the value than what is actually important. Do you concur Marc-André?

Just to avoid any confusion. The .launch file itself is pretty formatted, it's just the value of the 
attribute org.eclipse.dsf.launch.MEMORY_BLOCKS that is not with this change.

Patch-set: 5